[ADD] adventurer: tentative quest name in margin
Maybe I should put that on the right ? It's not a very important
information like the ALLIANCE and HORDE shields.

This is a problem when quest text starts at the last line of a
a page, sine the quest name then lies below the page number and
the hearthstone mark.