[ADD] adventurer: add new fetching mechanism for translations
I now use the TrinityCore database to fetch the information I need
when these are broadcast_texts or npcs so I can use ids and fetch
the translations for later when (if) someone ever translates this.

This still needs to be extended to quests however, as they are still
fetching the text from wowpedia. Wowpedia is great because it has
description, progress and completion (TrinityCore is missing a lot
of progress and completion texts), but wowpedia is not translated.